+++ playerbot/strategy/actions/MoveToRpgTargetAction.cpp
@@ -87,8 +87,8 @@ bool MoveToRpgTargetAction::isUseful()
     return context->GetValue<ObjectGuid>("rpg target")->Get()
         && (!context->GetValue<TravelTarget*>("travel target")->Get()->isTraveling() || !ChooseRpgTargetAction::isFollowValid(bot, context->GetValue<TravelTarget*>("travel target")->Get()->getLocation()))
         && AI_VALUE2(float, "distance", "rpg target") > sPlayerbotAIConfig.followDistance
-        && AI_VALUE2(uint8, "health", "self target") > sPlayerbotAIConfig.almostFullHealth
-        && (!AI_VALUE2(uint8, "mana", "self target") || AI_VALUE2(uint8, "mana", "self target") > sPlayerbotAIConfig.mediumMana)
+        && AI_VALUE2(float, "health", "self target") > sPlayerbotAIConfig.almostFullHealth
+        && (!AI_VALUE2(float, "mana", "self target") || AI_VALUE2(float, "mana", "self target") > sPlayerbotAIConfig.mediumMana)
         && !sServerFacade.IsInCombat(bot, true)
         && ChooseRpgTargetAction::isFollowValid(bot, wo);
 }
