+++ playerbot/strategy/generic/CastTimeStrategy.cpp
@@ -11,7 +11,7 @@ float CastTimeMultiplier::GetValue(Action* action)
 {
     if (action == NULL) return 1.0f;
 
-    uint8 targetHealth = AI_VALUE2(uint8, "health", "current target");
+    float targetHealth = AI_VALUE2(float, "health", "current target");
     string name = action->GetName();
 
     if (action->GetTarget() != AI_VALUE(Unit*, "current target"))
