+++ playerbot/strategy/hunter/HunterTriggers.cpp
@@ -8,7 +8,7 @@ using namespace ai;
 bool HunterNoStingsActiveTrigger::IsActive()
 {
 	Unit* target = AI_VALUE(Unit*, "current target");
-    return target && AI_VALUE2(uint8, "health", "current target") > 40 &&
+    return target && AI_VALUE2(float, "health", "current target") > 40 &&
         !ai->HasAura("serpent sting", target) &&
         !ai->HasAura("scorpid sting", target) &&
         !ai->HasAura("viper sting", target);
@@ -22,7 +22,7 @@ bool HuntersPetDeadTrigger::IsActive()
 bool HuntersPetLowHealthTrigger::IsActive()
 {
     Unit* pet = AI_VALUE(Unit*, "pet target");
-    return pet && AI_VALUE2(uint8, "health", "pet target") < 40 &&
+    return pet && AI_VALUE2(float, "health", "pet target") < 40 &&
         !AI_VALUE2(bool, "dead", "pet target") && !AI_VALUE2(bool, "mounted", "self target");
 }
